+++ versao.py
@@ -4,4 +4,4 @@ from typing import Final
 
 __all__ = ["VERSAO_APLICATIVO"]
 
-VERSAO_APLICATIVO: Final = "0.6.0"
+VERSAO_APLICATIVO: Final = "0.8.2"